+++ src/main/java/com/kozak/mybookshop/service/BookService.java
@@ -9,5 +9,5 @@ public interface BookService {
 
     List<BookDto> findAll();
 
-    BookDto getById(int id);
+    BookDto getById(Long id);
 }
